pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cairn"
version = "0.1.0"
description = "Utility functions for loading, querying, and analysing Zarr arrays of genomic data."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Tristan Dennis; tristanpwdennis@gmail.com" }
]
requires-python = ">=3.10,<3.12"
dependencies = [
  "scikit-allel",
  "anjl",
  "zarr",
  "xarray",
  "plotly",
  "seaborn",
  "numpy",
  "pandas",
  "google",
  "gcsfs",
  "scipy",
  "tqdm",
  "scikit-bio"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
  "mypy",
  "pre-commit"
]

[project.urls]
Homepage = "https://github.com/tristanpwdennis/cairn"
Issues = "https://github.com/tristanpwdennis/cairn/issues"

# Tool configs

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C90", "B", "A"]
ignore = ["E501"]  # long lines allowed since Black handles formatting
target-version = "py311"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
